requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.2.11
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.7.0
